+++ Core/src/Circle.hxx
@@ -13,8 +13,8 @@ public:
 	virtual void BeginShape() override;
 	virtual void EndShape() override;
 	virtual void AddVertex(const Vertex2& vertex) override;
-	virtual void RemoveVertex(int index) override;
-	virtual void ReplaceVertex(int index, Vertex2 newVertex) override;
+	virtual void RemoveVertex(size_t index) override;
+	virtual void ReplaceVertex(size_t index, Vertex2 newVertex) override;
 	virtual bool IsFinished() const override;
 	virtual ShapeType GetType() const override;
